requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Migrate==2.6.0
Flask-RESTful==0.3.9
flask-restplus==0.13.0
Flask-Script==2.0.5
Flask-SQLAlchemy==3.0.2
Flask-Testing==0.8.1
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.0
Mako==1.2.3
MarkupSafe==2.1.1
pycparser==2.18
PyJWT==2.6.0
pyrsistent==0.19.2
python-dateutil==2.7.0
python-editor==1.0.3
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.43
Werkzeug==2.2.2
